change frequencies to doubles -- all tests passing

+++ db-ext/course-reco/src/main/java/lernt/CandidateDecider.java
@@ -5,6 +5,10 @@ import org.neo4j.graphdb.*;
 import java.util.*;
 import static java.util.stream.Collectors.toMap;
 
+import org.apache.commons.math3.distribution.NormalDistribution;
+import org.apache.commons.math3.distribution.RealDistribution;
+import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
+
 
 /**
  * Given a course and a possible prereq, decides whether the prereq
@@ -56,13 +60,13 @@ public class CandidateDecider
             throws Exception
     {
         // Get all incoming courses and their relative frequencies, and calculate the total of incoming
-        Map<Node, Integer> freqs = getIncomingFrequencies();
-        int totalIncoming = freqs.values().stream().reduce(0, Integer::sum);
+        Map<Node, Double> freqs = getIncomingFrequencies();
+        double totalIncoming = freqs.values().stream().reduce(0.0, Double::sum);
 
 
         // Sort from highest to lower frequency
         // This is necessary when later we check for duplicates
-        Map<Node, Integer> sorted = freqs
+        Map<Node, Double> sorted = freqs
                 .entrySet()
                 .stream()
                 .sorted(Collections.reverseOrder(Map.Entry.comparingByValue()))
@@ -71,9 +75,9 @@ public class CandidateDecider
 
         Set<Course> candidateSet = new HashSet<>();
 
-        for (Map.Entry<Node, Integer> entry : sorted.entrySet())
+        for (Map.Entry<Node, Double> entry : sorted.entrySet())
         {
-            int currentFrequency = entry.getValue();
+            double currentFrequency = entry.getValue();
             if (!passesMinimumThreshold(currentFrequency, totalIncoming)) {
                 continue;
             }
@@ -110,12 +114,12 @@ public class CandidateDecider
      * This is due to the fact that our model creates multiple relationships between nodes to signify frequency
      * @return  The map of frequencies and nodes
      */
-    private Map<Node, Integer> getIncomingFrequencies()
+    private Map<Node, Double> getIncomingFrequencies()
     {
         RelationshipType prereqRelType = RelationshipType.withName(config.getPrereqLabelName());
         Iterable<Relationship> rels = current.getNode().getRelationships(prereqRelType, Direction.INCOMING);
 
-        Map<Node, Integer> map = new HashMap<>();
+        Map<Node, Double> map = new HashMap<>();
 
         for (Relationship rel : rels) {
             Node otherNode = rel.getStartNode();
@@ -124,12 +128,12 @@ public class CandidateDecider
             }
 
             // Add frequency or update it as needed
-            Integer freq = map.get(otherNode);
+            Double freq = map.get(otherNode);
             if (freq == null) {
-                map.put(otherNode, 1);
+                map.put(otherNode, 1.0);
             }
             else {
-                map.put(otherNode, freq + 1);
+                map.put(otherNode, freq + 1.0);
             }
         }
 
@@ -144,14 +148,14 @@ public class CandidateDecider
      * @return
      * @throws Exception If division by zero or negative number
      */
-    private boolean passesMinimumThreshold(int courseFrequency, int totalFrequency)
+    private boolean passesMinimumThreshold(double courseFrequency, double totalFrequency)
             throws Exception
     {
         if (totalFrequency <= 0) {
             throw new Exception("totalFrequency is not positive: " + totalFrequency);
         }
 
-        return (double) courseFrequency / (double) totalFrequency > this.frequencyThreshold;
+        return courseFrequency / totalFrequency > this.frequencyThreshold;
     }
